las naves ya disparan correctamente

+++ Source/GALAGA_USFX_LAB01/NaveEnemigaEspia.h
@@ -14,8 +14,8 @@ class GALAGA_USFX_LAB01_API ANaveEnemigaEspia : public ANaveEnemiga
 {
 	GENERATED_BODY()
 private:
-	int TiempoDeDisparo;
-	int Intervalo;
+	float TiempoDeDisparo;
+	float Intervalo;
 public:
 	ANaveEnemigaEspia();
 public:
